Code/requirements.txt
python3-opencv
numpy
pyserial
pupil-apriltags
picamera2